Added /health and /tokens
